move pkg/restore's backup extractor to pkg/archive
